Update View.java
Added JMenu items and bar incase the user wants to play again, undo a move or quit the game. Added an actionlistener to every tile in board so that the user can click where he wants the plants to be insteam or typing coordinates and also to all the buttons in jframe and menu items.